+++ CHANGELOG.md
@@ -67,6 +67,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (evm) [#169](https://github.com/EscanBE/evermint/pull/169) Support 256 blocks for `GetHashFn`
 - (app) [#177](https://github.com/EscanBE/evermint/pull/177) Fix staking hooks not available in some keepers
 - (evm) [#184](https://github.com/EscanBE/evermint/pull/184) Warmup coinbase + custom-precompiled-contracts and fix some tests
+- (cpc) [#186](https://github.com/EscanBE/evermint/pull/186) Correct genesis import/export for `x/cpc`
 
 ### API Breaking